Remove email address
The damage is likely already done but this is one less place my email may be out in the open to be scraped by spammers. There are other ways of reaching me like Twitter and anyone looking at commit history should be able to pull it out if they really need me.

This change likely won't make it to Docker Hub where it probably matters or at least not any time soon.