Create a build for PHP 8.2, likely very late.
Fixes #27

Fix README issues, w0rd-driven doesn't work as a user name now. Push wouldn't work because the local version w0rd-driven creates as w0rddriven. It's weird, Docker, deal with it.

I think it likely does create the local dash version but a push doesn't recognize it. Better to just change the name because since it's built, it just remaps to the new name instantly.